Delete Process

+++ nodes/nodeImplementations/deleteProcedure/DeleteProcess.java
@@ -6,12 +6,14 @@ import projects.cbrenet.nodes.messages.SDNMessage.DeleteMessage;
 import projects.cbrenet.nodes.messages.deletePhaseMessages.DeleteBaseMessage;
 import projects.cbrenet.nodes.messages.deletePhaseMessages.DeleteConfirmMessage;
 import projects.cbrenet.nodes.messages.deletePhaseMessages.DeletePrepareMessage;
-import projects.cbrenet.nodes.messages.deletePhaseMessages.FinishDeleteMessage;
+import projects.cbrenet.nodes.messages.deletePhaseMessages.DeleteFinishMessage;
+import projects.cbrenet.nodes.nodeImplementations.AuxiliaryNode;
 import projects.cbrenet.nodes.nodeImplementations.AuxiliaryNodeMessageQueueLayer;
 import projects.cbrenet.nodes.nodeImplementations.CounterBasedBSTLayer;
 import projects.cbrenet.nodes.nodeImplementations.MessageSendLayer;
 import projects.cbrenet.nodes.routeEntry.SendEntry;
 import sinalgo.nodes.Node;
+import sinalgo.nodes.messages.Message;
 import sinalgo.tools.Tools;
 
 import java.util.List;
@@ -37,15 +39,11 @@ public class DeleteProcess {
 
 
     public void startDelete(SendEntry deleteEntry, Node node, int largeId, int helpedId){
-
-
         // when node is node AN, node.id should equal to helpedId
         assert !(node instanceof CounterBasedBSTLayer) || node.ID == helpedId;
 
-
-        if(!deleteEntry.isDeleteFlag()){
-            Tools.warning("Want to delete, but can not now, " +
-                    "have to wait one neighbor is deleting.");
+        // if deleting, means no need to send DeletePrepareMessage
+        if(deleteEntry.isDeletingFlagOfItSelf()){
             return;
         }
 
@@ -57,7 +55,8 @@ public class DeleteProcess {
         deleteEntry.setDeletePrepareMessage(deletePrepareMessage);
         deleteEntry.initGotMap();
 
-        if(deleteEntry.cheakNeighborDeleting()){
+        if(deleteEntry.checkNeighborDeleting()){
+            deleteEntry.setDeletingFlagOfItSelf(true);
             for(int targetId : egoIdTargetList){
                 boolean upward = false;
                 if(deleteEntry.getRelationShipTo(targetId) == 'p'){
@@ -92,9 +91,9 @@ public class DeleteProcess {
     }
 
 
-    public void deleteConfirm(SendEntry deleteEntry, int helpedId, Node node){
+    public void neighborDeleteRequestConfirm(SendEntry deleteEntry, int helpedId, Node node){
         /**
-         *@description  Delete Confirm Phase
+         *@description Send DeleteConfirmMessage to the deleting one.
          *@parameters  [sendEntry, deletePrepareMessage, helpedId]
          *              helpedId: In AN, should be helpedId
          *                        In BST, should be BST' id, which is node.ID = helpedId
@@ -108,20 +107,26 @@ public class DeleteProcess {
         for(DeletePrepareMessage deletePrepareMessage : messageList){
             int largeId = deletePrepareMessage.getLargeId();
             int deleteTarget = deletePrepareMessage.getDeleteTarget();
-            // 等待
+
+            if(deleteTarget == helpedId){
+                // means this is its DeletePrepareMessage
+                continue;
+            }
+
+            // todo 完善部分结点不能回复 DeleteConfirmMessage时造成的问题
             DeleteConfirmMessage confirmMessage = new DeleteConfirmMessage(largeId, deleteTarget, helpedId);
+
+            deleteEntry.setDeletingFlag(deleteTarget);
 
             this.sendDeleteBaseMessage(confirmMessage, largeId, deleteEntry, deleteTarget, helpedId, node);
         }
     }
 
-    public boolean receiveDeleteConfirmMessage(SendEntry sendEntry, DeleteConfirmMessage deleteConfirmMessage, int helpedId){
-        int src = deleteConfirmMessage.getSrcId();
-        int largeId = deleteConfirmMessage.getLargeId();
+    public boolean receiveDeleteConfirmMessage(SendEntry sendEntry, DeleteConfirmMessage deleteConfirmMessage){
         return sendEntry.setDeleteConfirmMessage(deleteConfirmMessage);
     }
 
-    public void sendFinishDeleteMessage(SendEntry sendEntry, int largeId, int helpedId, Node node){
+    public void sendDeleteFinishMessage(SendEntry sendEntry, int largeId, int helpedId, Node node){
         List<Integer> ids = sendEntry.getAllSendIds();
 
 
@@ -130,10 +135,10 @@ public class DeleteProcess {
             DeleteMessage correspondingMessage = sendEntry.getCorrespondingDeleteMessage();
             int auxiliaryId = correspondingMessage.getAuxiliaryNodeId();
 
-            FinishDeleteMessage finishDeleteMessage = new FinishDeleteMessage(largeId, helpedId, helpedId, auxiliaryId);
+            DeleteFinishMessage deleteFinishMessage = new DeleteFinishMessage(largeId, helpedId, helpedId, auxiliaryId);
             for(int dstId : ids)
             {
-                this.sendDeleteBaseMessage(finishDeleteMessage, largeId, sendEntry, dstId, helpedId, node);
+                this.sendDeleteBaseMessage(deleteFinishMessage, largeId, sendEntry, dstId, helpedId, node);
             }
 
         }
@@ -146,23 +151,71 @@ public class DeleteProcess {
                 int sendIdOfNeighbor1 = sendEntry.getSendIdOf(neighbor1);
                 int sendIdOfNeighbor2 = sendEntry.getSendIdOf(neighbor2);
 
-                FinishDeleteMessage finishDeleteMessageTo2 = new FinishDeleteMessage(largeId, helpedId, neighbor1, sendIdOfNeighbor1);
-                FinishDeleteMessage finishDeleteMessageTo1 = new FinishDeleteMessage(largeId, helpedId, neighbor2, sendIdOfNeighbor2);
+                DeleteFinishMessage deleteFinishMessageTo2 = new DeleteFinishMessage(largeId, helpedId, neighbor1, sendIdOfNeighbor1);
+                DeleteFinishMessage deleteFinishMessageTo1 = new DeleteFinishMessage(largeId, helpedId, neighbor2, sendIdOfNeighbor2);
 
-                this.sendDeleteBaseMessage(finishDeleteMessageTo2, largeId, sendEntry, neighbor2, helpedId, node);
-                this.sendDeleteBaseMessage(finishDeleteMessageTo1, largeId, sendEntry, neighbor1, helpedId, node);
+                this.sendDeleteBaseMessage(deleteFinishMessageTo2, largeId, sendEntry, neighbor2, helpedId, node);
+                this.sendDeleteBaseMessage(deleteFinishMessageTo1, largeId, sendEntry, neighbor1, helpedId, node);
 
 
             }
             else if(ids.size() == 1){
                 // no need to create any link
                 int neighbor1 = ids.get(0);
-                FinishDeleteMessage finishDeleteMessageTo1 = new FinishDeleteMessage(largeId, helpedId, -1, -1);
-                this.sendDeleteBaseMessage(finishDeleteMessageTo1, largeId, sendEntry, neighbor1, helpedId, node);
+                DeleteFinishMessage deleteFinishMessageTo1 = new DeleteFinishMessage(largeId, helpedId, -1, -1);
+                this.sendDeleteBaseMessage(deleteFinishMessageTo1, largeId, sendEntry, neighbor1, helpedId, node);
             }
             else{
                 Tools.warning("Very Interesting situation， no node connect to it!");
             }
+        }
+
+    }
+
+
+    public void executeDeleteFinishMessage(SendEntry sendEntry, Node node, DeleteFinishMessage deleteFinishMessage, int helpedId){
+        int deleteTarget = deleteFinishMessage.getDeleteTarget();
+        int largeId = deleteFinishMessage.getLargeId();
+
+        int egoTreeId = deleteFinishMessage.getEgoTreeId();
+        int trueId = deleteFinishMessage.getTrueId();
+
+        char relation = sendEntry.getRelationShipTo(deleteTarget);
+
+        switch (relation){
+            case 'p':
+                if(node instanceof MessageSendLayer){
+                    MessageSendLayer messageSendLayer = (MessageSendLayer) node;
+                    messageSendLayer.changeParentTo(largeId, egoTreeId, trueId);
+                }
+                else if(node instanceof AuxiliaryNode){
+                    AuxiliaryNode auxiliaryNode = (AuxiliaryNode) node;
+                    auxiliaryNode.changeParentTo(largeId, egoTreeId, trueId, helpedId);
+                }
+                break;
+            case 'l':
+                if(node instanceof MessageSendLayer){
+                    MessageSendLayer messageSendLayer = (MessageSendLayer) node;
+                    messageSendLayer.changeLeftChildTo(largeId, egoTreeId, trueId);
+                }
+                else if(node instanceof AuxiliaryNode){
+                    AuxiliaryNode auxiliaryNode = (AuxiliaryNode) node;
+                    auxiliaryNode.changeLeftChildTo(largeId, egoTreeId, trueId, helpedId);
+                }
+                break;
+            case 'r':
+                if(node instanceof MessageSendLayer){
+                    MessageSendLayer messageSendLayer = (MessageSendLayer) node;
+                    messageSendLayer.changeRightChildTo(largeId, egoTreeId, trueId);
+                }
+                else if(node instanceof AuxiliaryNode){
+                    AuxiliaryNode auxiliaryNode = (AuxiliaryNode) node;
+                    auxiliaryNode.changeRightChildTo(largeId, egoTreeId, trueId, helpedId);
+                }
+                break;
+            default:
+                Tools.warning("When executing DeleteFinishMessage, can not get relation");
+                break;
         }
 
     }
